requirements.txt
torch==2.6.0
torchvision==0.21.0
click==8.2.1
einops==0.8.1
hydra-core==1.3.2
hydra-joblib-launcher==1.2.0
jupyter==1.1.1
lightning==2.5.2
pot==0.9.5
rich==14.0.0
scikit-learn==1.7.0
seaborn==0.13.2
torch-ema==0.3.0
transformers==4.52.4
torch-fidelity==0.3.0
pqm==0.6.3
comet_ml==3.49.12